Override the variable display-buffer-overriding-action
Some functions, like `pop-to-buffer-same-window' pass an action that inhibits
the appearance of new buffers in the same window. In such cases, buffers will
silently fail to appear. An example of such a function that uses
`pop-to-buffer-same-window' internally is the `scratch-buffer' interactive
function. Enabling `current-window-only-mode' and running said function will
cause Emacs to hang indefinitely. Due to the design of the buffer display
framework, `display-buffer-overriding-action' is the most prioritized of all
variables. Despite the  documentation's warnings not to ever set this variable
in Lisp programs, there is really no way to avoid this problem without
overriding it.